nits: address review comments
